+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [unreleased]
 
+_Use self defined objects instead of jsonrpc2-objects and jsonrpc2-pyclient._
+
+### Changed
+
+- Substitute `jsonrpc2-objects` and `jsonrpc2-pyclient` by self written objects.
+
 ### Added
 
 - Add __future__.annotations to all files, which need it for annotations for Python 3.7/3.8.